Updates to v2 API
