--- dev/td-modules/base_hueControl/requirements.txt
@@ -1,1 +1,0 @@
-phue
